Ambitious reimagining of the wizard model with a new wizard config
