+++ src/commands/tokenAdd.js
@@ -1,4 +1,5 @@
-const { WizardAddTokenRule } = require("../wizard/add-token-rule");
+const { AddTokenRuleWizardConfig } = require("../wizard/add-token-rule");
+const { Wizard } = require("../wizard/wizard");
 const { globalUserWizards } = require("../wizard/wizard");
 
 ///
@@ -11,7 +12,15 @@ async function starryCommandTokenAdd(interaction, client) {
 
 	// TODO: this is where we'll want to do a filter/map deal to remove all entries that have a {wizard}.createdAt that's > some amount, like 6 minutes
 
-	let addTokenRuleWizard = new WizardAddTokenRule(interaction.guildId, interaction.channelId, userId, client)
+	// let addTokenRuleWizard = new WizardAddTokenRule(interaction.guildId, interaction.channelId, userId, client)
+	let addTokenRuleWizard = new Wizard(
+		AddTokenRuleWizardConfig,
+		interaction.guildId,
+		interaction.channelId,
+		userId,
+		client
+	)
+
 	// Begin the wizard by calling the begin function on the first step
 	const msg = await addTokenRuleWizard.currentStep.beginFn({interaction});
